Implement Test Optimization Problem and Baseline Model Prototype
The test problem is approximating the Taylor expansion of sin(x)
on [-pi, pi] as a vector of coefficients. The baseline model
randomly mutates its guess and steps to the new solution if it
scores better.

I'll improve this model some more, then try my hand at some
advanced models that use pools of guesses.