Complete Greedy Walk Implementation (Baseline Model)
The initial 2018 implementation of the greedy walk algorithm
achieved a least sum of squares of .209 on [-pi, pi], then .028
if I removed that `.pop()` to remove coefficients. On [-2pi, 2pi],
it would remove (or without `.pop()`, vanish) all coefficients and
stay at the theoretical max cost 0.5.

This new implementation achieves <.001 on [-pi, pi] and .047 on
[-2pi, 2pi]. Solid stuff.